chore: added tooltips to show which file type is going to be uploaded
